Feat/roles (#25)

+++ backend/seeders/20250208124437-super-admin.js
@@ -17,10 +17,6 @@ export default {
   },
 
   down: async (queryInterface, Sequelize) => {
-    await queryInterface.bulkDelete(
-      'users',
-      { useremail: 'superadmin@example.com' },
-      {}
-    )
+    await queryInterface.bulkDelete('users', { username: 'superadmin' }, {})
   },
 }
